missing type cast

+++ media/libstagefright/ACodecBufferChannel.cpp
@@ -285,7 +285,7 @@ status_t ACodecBufferChannel::attachBuffer(
             size_t copyLength = std::min(size_t(view.capacity()), buffer->capacity());
             ALOGV_IF(view.capacity() > buffer->capacity(),
                     "view.capacity() = %zu, buffer->capacity() = %zu",
-                    view.capacity(), buffer->capacity());
+                    size_t(view.capacity()), buffer->capacity());
             memcpy(buffer->base(), view.data(), copyLength);
             buffer->setRange(0, copyLength);
             break;
